refactor(deps): restructure into checks/
