+++ avd_docs/kubernetes/general/AVD-KSV-0004/docs.md
@@ -1,0 +1,13 @@
+
+Security best practices require containers to run with minimal required capabilities.
+
+### Impact
+<!-- Add Impact here -->
+
+<!-- DO NOT CHANGE -->
+{{ remediationActions }}
+
+### Links
+- https://kubesec.io/basics/containers-securitycontext-capabilities-drop-index-all/
+
+
