+++ avd_docs/kubernetes/general/AVD-KSV-0034/docs.md
@@ -1,0 +1,10 @@
+
+Container images must not start with an empty prefix or a defined public registry domain.
+
+### Impact
+<!-- Add Impact here -->
+
+<!-- DO NOT CHANGE -->
+{{ remediationActions }}
+
+
