+++ avd_docs/kubernetes/general/AVD-KSV-0040/docs.md
@@ -1,0 +1,13 @@
+
+ensure resource quota policy has configure in order to limit aggregate resource usage within namespace
+
+### Impact
+<!-- Add Impact here -->
+
+<!-- DO NOT CHANGE -->
+{{ remediationActions }}
+
+### Links
+- https://kubernetes.io/docs/tasks/administer-cluster/manage-resources/quota-memory-cpu-namespace/
+
+
