+++ internal/rules/register.go
@@ -8,7 +8,7 @@ import (
 	"github.com/aquasecurity/defsec/pkg/framework"
 	"github.com/aquasecurity/defsec/pkg/scan"
 	dftypes "github.com/aquasecurity/defsec/pkg/types"
-	"github.com/aquasecurity/trivy-policies/rules/specs"
+	"github.com/aquasecurity/trivy-policies/specs"
 
 	"github.com/aquasecurity/trivy-iac/pkg/types"
 )
