+++ pkg/rules/rules.go
@@ -1,78 +1,78 @@
 package rules
 
 import (
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/accessanalyzer"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/apigateway"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/athena"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/cloudfront"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/cloudtrail"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/cloudwatch"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/codebuild"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/config"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/documentdb"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/dynamodb"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/ec2"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/ecr"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/ecs"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/efs"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/eks"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/elasticache"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/elasticsearch"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/elb"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/emr"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/iam"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/kinesis"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/kms"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/lambda"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/mq"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/msk"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/neptune"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/rds"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/redshift"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/s3"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/sam"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/sns"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/sqs"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/ssm"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/aws/workspaces"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/appservice"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/authorization"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/compute"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/container"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/database"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/datafactory"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/datalake"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/keyvault"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/monitor"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/network"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/securitycenter"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/storage"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/azure/synapse"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/cloudstack/compute"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/digitalocean/compute"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/digitalocean/spaces"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/github/actions"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/github/branch_protections"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/github/repositories"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/google/bigquery"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/google/compute"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/google/dns"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/google/gke"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/google/iam"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/google/kms"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/google/sql"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/google/storage"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/nifcloud/computing"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/nifcloud/dns"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/nifcloud/nas"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/nifcloud/network"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/nifcloud/rdb"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/nifcloud/sslcertificate"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/openstack/compute"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/openstack/networking"
+	_ "github.com/aquasecurity/trivy-policies/checks/cloud/oracle/compute"
+	_ "github.com/aquasecurity/trivy-policies/checks/kubernetes/network"
 	trules "github.com/aquasecurity/trivy-policies/pkg/rules"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/accessanalyzer"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/apigateway"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/athena"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/cloudfront"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/cloudtrail"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/cloudwatch"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/codebuild"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/config"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/documentdb"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/dynamodb"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/ec2"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/ecr"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/ecs"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/efs"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/eks"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/elasticache"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/elasticsearch"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/elb"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/emr"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/iam"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/kinesis"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/kms"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/lambda"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/mq"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/msk"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/neptune"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/rds"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/redshift"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/s3"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/sam"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/sns"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/sqs"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/ssm"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/aws/workspaces"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/appservice"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/authorization"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/compute"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/container"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/database"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/datafactory"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/datalake"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/keyvault"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/monitor"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/network"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/securitycenter"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/storage"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/azure/synapse"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/cloudstack/compute"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/digitalocean/compute"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/digitalocean/spaces"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/github/actions"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/github/branch_protections"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/github/repositories"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/google/bigquery"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/google/compute"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/google/dns"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/google/gke"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/google/iam"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/google/kms"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/google/sql"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/google/storage"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/nifcloud/computing"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/nifcloud/dns"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/nifcloud/nas"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/nifcloud/network"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/nifcloud/rdb"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/nifcloud/sslcertificate"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/openstack/compute"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/openstack/networking"
-	_ "github.com/aquasecurity/trivy-policies/rules/cloud/policies/oracle/compute"
-	_ "github.com/aquasecurity/trivy-policies/rules/kubernetes/network"
 )
 
 func init() {
